Fix fp16 NaN: GeM in fp32, attention in fp32, clamp augmentations, lower LR
Root cause: fp16 overflow in GeM pooling (pow(p) with learnable p)
and window self-attention (QK^T dot products). Once NaN enters weights,
model never recovers even with grad skip.

Fixes:
- GeM: compute pow() in fp32, clamp p to [1, 6]
- WindowAttention: QK^T and softmax in fp32
- Augmentation: clamp values to [-10, 10] after all transforms
- Default LR: 3e-4 → 1e-4 (less aggressive with AMP)

+++ model/bottleneck.py
@@ -105,14 +105,15 @@ class WindowAttention(nn.Module):
         qkv = qkv.permute(2, 0, 3, 1, 4)  # (3, Bw, heads, N, head_dim)
         q, k, v = qkv.unbind(0)
 
-        attn = (q @ k.transpose(-2, -1)) * self.scale
+        # Compute attention in fp32 to prevent overflow in fp16
+        attn = (q.float() @ k.float().transpose(-2, -1)) * self.scale
 
         # Add relative position bias
         rel_bias = self.rel_pos_bias[self.relative_position_index.view(-1)].view(
             N, N, -1).permute(2, 0, 1)
-        attn = attn + rel_bias.unsqueeze(0)
+        attn = attn + rel_bias.unsqueeze(0).float()
 
-        attn = attn.softmax(dim=-1)
+        attn = attn.softmax(dim=-1).to(v.dtype)
 
         out = (attn @ v).transpose(1, 2).reshape(Bw, N, C)
         out = self.proj(out)
